% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHandler.r
\name{aceHotkeyHandler}
\alias{aceHotkeyHandler}
\title{Add an handler to a hotkey in an aceEditor component}
\usage{
aceHotkeyHandler(id, fun, ..., app = getApp(),
  if.handler.exists = c("replace", "add", "skip")[1],
  session = getAppSession(app))
}
\arguments{
\item{id}{name of the button}

\item{fun}{function that will be called if button is pressed. The function will be called with the following arguments:

 keyId: the id assigned to the hotkey
 editorId: the id of the aceEditor widget
 selection: if a text is selected, this selection
 text: the text of the aceEditor widget
 cursor: a list with the current cursor position:
         row and column with index starting with 0
 session: the current session object}

\item{...}{extra arguments that will be passed to fun when the event is triggered.}
}
\description{
Add an handler to a hotkey in an aceEditor component
}

