% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addButtonHandler}
\alias{addButtonHandler}
\title{Add an handler to a button}
\usage{
addButtonHandler(id, fun, ..., app = getApp())
}
\arguments{
\item{id}{name of the button}

\item{fun}{function that will be called if button is pressed. The function will be called with the arguments: 'id', 'value' and 'session'. One can assign the same handler functions to several buttons.}
}
\description{
Add an handler to a button
}

