% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHandler.r
\name{appInitHandler}
\alias{appInitHandler}
\title{Set a function that will be called when a new session of
an app is initialized.}
\usage{
appInitHandler(initHandler, app = getApp())
}
\description{
@param initHandler a function that takes parameters
   session, input, output and app. It will be called from
   the app$server function whenever a new session is created.
   It allows to initialize session specific variables and e.g.
   store them within the app object. The passed app object is already
   the local copy created for the new session.
}

