% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{bottomScript}
\alias{bottomScript}
\title{If app is not running, mark script to be added at the bottom and return NULL
If app is already running return script directly}
\usage{
bottomScript(..., app = getApp())
}
\description{
If app is not running, mark script to be added at the bottom and return NULL
If app is already running return script directly
}

