% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{buttonHandler}
\alias{buttonHandler}
\title{A more efficient version of button handler via global eventId handler}
\usage{
buttonHandler(id, fun, ..., eventId = "buttonHandlerEvent",
  jscript = buttonHandlerJS(eventId), app = getApp(),
  no.authentication.required = FALSE)
}
\arguments{
\item{id}{buttonId}

\item{fun}{the handler fun that will be called when the button is clicked}

\item{...}{additional arguments passed to the handler fun}
}
\description{
A more efficient version of button handler via global eventId handler
}

