% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{callJS}
\alias{callJS}
\title{Call a javascript function or method with R arguments}
\usage{
callJS(.fun, ..., .args = NULL, .app = getApp())
}
\description{
Call a javascript function or method with R arguments
}

