% Generated by roxygen2 (4.0.1): do not edit by hand
\name{changeHandler}
\alias{changeHandler}
\title{Add an handler to an input that is called when the input value changes}
\usage{
changeHandler(id, fun, ..., app = getApp(), on.create = FALSE)
}
\arguments{
\item{id}{name of the input element}

\item{fun}{function that will be called if the input value changes. The function will be called with the arguments: 'id', 'value' and 'session'. One can assign the same handler functions to several input elements.}
}
\description{
Add an handler to an input that is called when the input value changes
}

