% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{classEventHandler}
\alias{classEventHandler}
\title{An event handler for objects with given class}
\usage{
classEventHandler(class, fun, event = "change", css.locator = "",
  inner.js.code = NULL, shiny.value.code = NULL, eventId = paste0(class,
  "_class_", event, "_event"), class.prefix = ".", ...)
}
\description{
An event handler for objects with given class
}

