% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directUpdate.r
\name{dsetUI}
\alias{dsetUI}
\title{Directly setUI , also works for hidden UI}
\usage{
dsetUI(id, ui, selector = paste0("#", id, collapse = ", "), app = getApp(),
  ...)
}
\description{
Directly setUI , also works for hidden UI
}

