% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{evalJS}
\alias{evalJS}
\title{Evaluate arbitrary java script code in the client's web browser}
\usage{
evalJS(js, ..., .args = list(...), app = getApp())
}
\description{
Evaluate arbitrary java script code in the client's web browser
}

