% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.r
\name{getApp}
\alias{getApp}
\title{get the current app object}
\usage{
getApp(session = NULL)
}
\description{
If the app is already running, gets by default the local app copy
corresponding to the current session
}

