% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventData.R
\name{getEventData}
\alias{getEventData}
\title{Get Event Data Table}
\usage{
getEventData(data = NULL, list = NULL)
}
\arguments{
\item{data}{Data frame object of the following columns: Data Table Name, Data File Path, Event Column, Expanded View Flag, Event Name, Event Start Column, Event Stop Column, Treatment Flag, Response Flag. More information in details.}

\item{list}{A list object of data frames annotated in the data object. The names of each data frame in the list should match the first column of the data object.}
}
\value{
Data frame.
}
\description{
Get Event Data Table
}
