% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.r
\name{getInputValue}
\alias{getInputValue}
\title{Get an input value from the current session}
\usage{
getInputValue(id, session = getAppSession(app), app = getApp())
}
\description{
Get an input value from the current session
}

