% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHandler.r
\name{hasButtonCounterIncreased}
\alias{hasButtonCounterIncreased}
\title{Checks whether a button has been pressed again (internal function)}
\usage{
hasButtonCounterIncreased(id, counter, app = getApp())
}
\description{
Checks whether a button has been pressed again (internal function)
}

