% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hasWidgetValueChanged}
\alias{hasWidgetValueChanged}
\title{Checks whether the value of an input item has been changed (internal function)}
\usage{
hasWidgetValueChanged(id, new.value, on.create = FALSE, app = getApp())
}
\description{
Checks whether the value of an input item has been changed (internal function)
}

