% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{idEventHandler}
\alias{idEventHandler}
\title{An event handler for an object with given id}
\usage{
idEventHandler(id, fun, event = "change", css.locator = "",
  inner.js.code = NULL, shiny.value.code = NULL, eventId = paste0(id,
  "_id_", event, "_event"), stop.propagation = FALSE, ...)
}
\description{
An event handler for an object with given id
}

