% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{imageClickHandler}
\alias{imageClickHandler}
\title{Handler for an image click}
\usage{
imageClickHandler(id, fun, ..., eventId = "imageClickEvent", app = getApp(),
  no.authentication.required = FALSE)
}
\arguments{
\item{id}{id of the HTML img object}

\item{fun}{the handler fun that will be called when the image is clicked}

\item{...}{additional arguments passed to the handler fun}
}
\description{
Handler for an image click
}

