% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{initialQueryDispatch}
\alias{initialQueryDispatch}
\title{Can be called inside initApp handler
fun is a function that gets an argument
query and can do some initial dispatch
depending on the query.
For some reason we need to use the
observer trick to get access to the
query object. This means dispatch
takes place after other commands in the
initApp handler.}
\usage{
initialQueryDispatch(fun, app = getApp(), ...)
}
\description{
Can be called inside initApp handler
fun is a function that gets an argument
query and can do some initial dispatch
depending on the query.
For some reason we need to use the
observer trick to get access to the
query object. This means dispatch
takes place after other commands in the
initApp handler.
}

