% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_setFunctions.r
\name{make.set.for.update.functions}
\alias{make.set.for.update.functions}
\title{Update an ",base.name," object for an shiny events app}
\usage{
make.set.for.update.functions(pkg = "shiny")
}
\description{
Similar to update",base.name," in the package ", pkg, ",
but no need to pass a session object.\n",
}

