% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{moveBottomScripts}
\alias{moveBottomScripts}
\title{Given a tag object, extract out any children of tags$head
and return them separate from the body.}
\usage{
moveBottomScripts(ui, reset.app = FALSE)
}
\description{
Given a tag object, extract out any children of tags$head
and return them separate from the body.
}

