% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{selectChangeHandler}
\alias{selectChangeHandler}
\title{Add an handler to an input or select that is called when the input value changes}
\usage{
selectChangeHandler(id, fun, ..., eventId = "selectChangeHandlerEvent",
  jscript = buttonHandlerJS(eventId), app = getApp())
}
\arguments{
\item{id}{name of the input element}

\item{fun}{function that will be called if the input value changes. The function will be called with the arguments: 'id', 'value' and 'session'. One can assign the same handler functions to several input elements.}

\item{...}{extra arguments that will be passed to fun when the event is triggered.}
}
\description{
Add an handler to an input or select that is called when the input value changes
}

