% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setDataTable}
\alias{setDataTable}
\title{Update an dataTableOutput object.}
\usage{
setDataTable(id, val, update.env = parent.frame(), app = getApp(), ...)
}
\description{
Can be used instead of renderDataTable.
Similar to updateDataTable but no need to provide session object
}

