% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{setHtmlAttribute}
\alias{setHtmlAttribute}
\title{Set attributes of HTML elements}
\usage{
setHtmlAttribute(id = NULL, attr, class = NULL,
  selector = paste0(c(sc("#", id), sc(".", class)), collapse = ", "),
  app = getApp())
}
\description{
Set attributes of HTML elements
}

