% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{setHtmlCSS}
\alias{setHtmlCSS}
\title{Set css style of HTML elements}
\usage{
setHtmlCSS(id = NULL, attr, class = NULL, selector = paste0(c(sc("#", id),
  sc(".", class)), collapse = ", "), app = getApp())
}
\description{
Set css style of HTML elements
}

