% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{setHtmlHide}
\alias{setHtmlHide}
\title{Hide HTML elements}
\usage{
setHtmlHide(id = NULL, class = NULL, display = "none",
  selector = paste0(c(sc("#", id), sc(".", class)), collapse = ", "))
}
\description{
Hide HTML elements
}

