% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{setHtmlShow}
\alias{setHtmlShow}
\title{Show HTML elements}
\usage{
setHtmlShow(id = NULL, class = NULL, display = "block",
  visibility = "visible", selector = paste0(c(sc("#", id), sc(".", class)),
  collapse = ", "))
}
\description{
Show HTML elements
}

