% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{setImage}
\alias{setImage}
\title{Update an output object. Can be used instead of renderImage}
\usage{
setImage(id, value, app = getApp(), ...)
}
\description{
Similar to updateImage but no need to provide session object
}

