% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{setPlot}
\alias{setPlot}
\title{update an plotOutput object. Can be used instead of renderPlot.}
\usage{
setPlot(id, expr, app = getApp(), update.env = parent.frame(),
  quoted = FALSE, ...)
}
\description{
Similar to updatePlot but no need to provide session object
}

