% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{setPrint}
\alias{setPrint}
\title{Update an textOutput object. Can be used instead of renderPrint}
\usage{
setPrint(id, expr, app = getApp(), ...)
}
\description{
Similar to updatePrint but no need to provide session object
}

