% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{setRHandsontable}
\alias{setRHandsontable}
\title{Update an RHandsontable object. Can be used instead of renderRHandsontable}
\usage{
setRHandsontable(id, value, app = getApp(), ...)
}
\description{
Similar to updateRHandsontable but no need to provide session object
}

