% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{setTable}
\alias{setTable}
\title{Update an tableOutput object. Can be used instead of renderTable}
\usage{
setTable(id, value, app = getApp(), ...)
}
\description{
Similar to updateTable but no need to provide session object
}

