% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{setText}
\alias{setText}
\title{Update an textOutput object. Can be used instead of renderText}
\usage{
setText(id, value, app = getApp(), ...)
}
\description{
Similar to updateText but no need to provide session object
}

