% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{setUI}
\alias{setUI}
\title{Update an uiOutput object. Can be used instead of renderUI}
\usage{
setUI(id, ui, app = getApp(), ...)
}
\description{
Similar to updateUI but no need to provide session object
}

