% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_events.R
\name{shiny_events}
\alias{shiny_events}
\title{Generic event implementation}
\usage{
shiny_events(app = basename(getwd()))
}
\arguments{
\item{app}{The name of the app. Defaults to the name of the script's containing folder.}
}
\value{
An environment variable containing: a GUID, the name of the app, entry() function and event() function.
}
\description{
Customize the output or add a new target file (beyond CSV, log or Database).
To do that, override the `event()` function after assigning it to a variable.
}
