% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_events.R
\name{shiny_events_to_csv}
\alias{shiny_events_to_csv}
\title{Record Shiny events on a CSV file}
\usage{
shiny_events_to_csv(app = basename(getwd()),
  filename = "shiny-events.csv")
}
\arguments{
\item{app}{The name of the app. Defaults to the name of the script's containing folder.}

\item{filename}{CSV file name to use. Defaults to "shiny-events.csv"}
}
\value{
An environment variable containing: a GUID, the name of the app, entry() function and event() function.
}
\description{
Record Shiny events on a CSV file
}
\examples{

file_name <- tempfile(fileext = ".csv")
tracker <- shiny_events_to_csv("example-app", file_name)
tracker$event("slider", 1)
read.csv(
  file_name,
  header = FALSE,
  col.names = c("guid", "app", "activity", "value", "datetime")
)
}
