% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_events.R
\name{shiny_events_to_log}
\alias{shiny_events_to_log}
\title{Record Shiny events on a LOG file}
\usage{
shiny_events_to_log(app = basename(getwd()),
  filename = "shiny-events.log")
}
\arguments{
\item{app}{The name of the app. Defaults to the name of the script's containing folder.}

\item{filename}{CSV file name to use. Defaults to "shiny-events.csv"}
}
\value{
An environment variable containing: a GUID, the name of the app, entry() function and event() function.
}
\description{
Record Shiny events on a LOG file
}
\examples{

file_name <- tempfile(fileext = ".csv")
tracker <- shiny_events_to_log("example-app", file_name)
tracker$event("slider", 1)
# event() function allows the log type to be modified
tracker$event("records-returned", 0, type = "WARN")
readLines(file_name)
}
