% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_events.R
\name{shiny_events_to_logger}
\alias{shiny_events_to_logger}
\title{Record Shiny events using the logger package}
\usage{
shiny_events_to_logger(app = basename(getwd()))
}
\arguments{
\item{app}{The name of the app. Defaults to the name of the script's containing folder.}
}
\value{
An environment variable containing: a GUID, the name of the app, entry() function and event() function.
}
\description{
Record Shiny events using the logger package
}
