% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{singletonBottomScript}
\alias{singletonBottomScript}
\title{If app is not running, mark script to be added at the bottom and return NULL
If app is already running return script directly}
\usage{
singletonBottomScript(..., app = getApp())
}
\description{
If app is not running, mark script to be added at the bottom and return NULL
If app is already running return script directly
}

