% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\name{svgClickHandler}
\alias{svgClickHandler}
\title{Handler for an image click}
\usage{
svgClickHandler(id, fun, ..., eventId = if (stop.propagation) "svgClickEvent"
  else "svgClickEventWithPropagation", class = "clickable_svg",
  app = getApp(), no.authentication.required = FALSE,
  stop.propagation = TRUE)
}
\arguments{
\item{id}{id of the HTML img object}

\item{fun}{the handler fun that will be called when the image is clicked}

\item{...}{additional arguments passed to the handler fun}
}
\description{
Handler for an image click
}

