% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelinePlot.R
\name{timelinePlot}
\alias{timelinePlot}
\title{Timeline plot}
\usage{
timelinePlot(
  data = NULL,
  event_col = NULL,
  event_type_col = NULL,
  start_col = NULL,
  stop_col = NULL,
  unit = "years",
  plotly = FALSE,
  title = "Timeline Plot",
  title_font = 16,
  x_font = 14,
  y_font = 14,
  na.rm = TRUE
)
}
\arguments{
\item{data}{Data frame object with columns detailing the specific event, event category, event start time and event stop time}

\item{event_col}{String. Column name of specific event name column. Defaults to first column.}

\item{event_type_col}{String. Column name of event category column. Defaults to second column.}

\item{start_col}{String. Column name of event start time column. Defaults to third column.}

\item{stop_col}{String. Column name of event stop time column. Defaults to fourth column.}

\item{unit}{String. Unit of time to label the x-axis.}

\item{plotly}{Boolean. If TRUE, function will return an interactive plotly object.}

\item{title}{String. Desired plot title.}

\item{title_font}{Numeric. Size of title font.}

\item{x_font}{Numeric. Size of font on the x-axis.}

\item{y_font}{Numeric. Size of font on the y-axis}

\item{na.rm}{Boolean. If TRUE, events with incomplete time data will be removed.}
}
\value{
plot
}
\description{
\code{timelinePlot()} is used to generate a patient timeline plot to view longitudinal data.
}
