% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateDataTable}
\alias{updateDataTable}
\title{Update an dataTableOutput object. Can be used instead of renderDataTable}
\usage{
updateDataTable(id, expr, update.env = parent.frame(), app = getApp())
}
\description{
Update an dataTableOutput object. Can be used instead of renderDataTable
}

