% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{updateDownloadHandler}
\alias{updateDownloadHandler}
\title{Shiny events version of downloadHandler}
\usage{
updateDownloadHandler(session = NULL, id, filename, content,
  contentType = NA, ..., app = getApp(session))
}
\arguments{
\item{id}{name of the downloadButton or downloadLink}

\item{filename}{A string of the filename, including extension, that the user's web browser should default to when downloading the file; or a function that returns such a string.}

\item{content}{A function that takes a single argument file that is a file path (string) of a nonexistent temp file, and writes the content to that file path.}

\item{contentType}{A string of the download's content type, for example "text/csv" or "image/png". If NULL or NA, the content type will be guessed based on the filename extension, or application/octet-stream if the extension is unknown.}
}
\description{
Shiny events version of downloadHandler
}

