% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateImage}
\alias{updateImage}
\title{Update an output object. Can be used instead of renderImage}
\usage{
updateImage(session = NULL, id, val, app = getApp(session), ...)
}
\description{
Update an output object. Can be used instead of renderImage
}

