% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updatePlot}
\alias{updatePlot}
\title{update an plotOutput object. Can be used instead of renderPlot.}
\usage{
updatePlot(id, expr, app = getApp(), update.env = parent.frame())
}
\description{
update an plotOutput object. Can be used instead of renderPlot.
}

