% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{updateRHandsontable}
\alias{updateRHandsontable}
\title{Update an RHandsontable object. Can be used instead of renderRHandsontable}
\usage{
updateRHandsontable(session = NULL, id, value, app = getApp(session), ...)
}
\description{
Update an RHandsontable object. Can be used instead of renderRHandsontable
}

