% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateText}
\alias{updateText}
\title{Update an textOutput object. Can be used instead of renderText}
\usage{
updateText(session = NULL, id, value, app = getApp(session), ...)
}
\description{
Update an textOutput object. Can be used instead of renderText
}

