% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateUI}
\alias{updateUI}
\title{Update an uiOutput object. Can be used instead of renderUI}
\usage{
updateUI(id, expr, update.env = parent.frame(), app = getApp())
}
\description{
Update an uiOutput object. Can be used instead of renderUI
}

