% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{viewApp}
\alias{viewApp}
\title{view shiny events app in RStudio viewer}
\usage{
viewApp(app = getApp(), ui = NULL, launch.browser = rstudio::viewer, ...)
}
\description{
view shiny events app in RStudio viewer
}

