% Generated by roxygen2 (4.0.1): do not edit by hand
\name{appInitHandler}
\alias{appInitHandler}
\title{Set a function that will be called when a new session of
an app is initialized.}
\usage{
appInitHandler(initHandler, app = getApp())
}
\arguments{
\item{initHandler}{a function that takes parameters
session, input, output and app. It will be called from
the app$server function whenever a new session is created.
It allows to initialize session specific variables and e.g.
store them within the app object. The passed app object is already
the local copy created for the new session.}
}
\description{
Set a function that will be called when a new session of
an app is initialized.
}

