% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHandler.r
\name{changeHandler}
\alias{changeHandler}
\title{Add an handler to an input that is called when the input value changes}
\usage{
changeHandler(id, fun, ..., app = getApp(), on.create = FALSE,
  if.handler.exists = c("replace", "add", "skip")[1],
  session = getAppSession(app), no.authentication.required = FALSE)
}
\arguments{
\item{id}{name of the input element}

\item{fun}{function that will be called if the input value changes. The function will be called with the arguments: 'id', 'value' and 'session'. One can assign the same handler functions to several input elements.}

\item{...}{extra arguments that will be passed to fun when the event is triggered.}
}
\description{
Add an handler to an input that is called when the input value changes
}

