% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHandler.r
\name{hasWidgetValueChanged}
\alias{hasWidgetValueChanged}
\title{Checks whether the value of an input item has been changed (internal function)}
\usage{
hasWidgetValueChanged(id, new.value, on.create = FALSE, app = getApp())
}
\description{
Checks whether the value of an input item has been changed (internal function)
}

