% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{setDataTable}
\alias{setDataTable}
\title{Update an dataTableOutput object.}
\usage{
setDataTable(id, value, app = getApp(), ...)
}
\description{
Can be used instead of renderDataTable.
Similar to updateDataTable but no need to provide session object
}

