% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setPlot}
\alias{setPlot}
\title{update an plotOutput object. Can be used instead of renderPlot.}
\usage{
setPlot(id, expr, app = getApp(), update.env = parent.frame(), ...)
}
\description{
Similar to updatePlot but no need to provide session object
}

