% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelinePlot.R
\name{timelinePlot}
\alias{timelinePlot}
\title{Timeline plot}
\usage{
timelinePlot(
  data = NULL,
  event_col = NULL,
  event_type_col = NULL,
  start_col = NULL,
  stop_col = NULL,
  unit = "years",
  plotly = FALSE,
  title = "Timeline Plot",
  title_font = 16,
  x_font = 14,
  y_font = 14,
  na.rm = TRUE
)
}
\arguments{
\item{data}{Data frame object with columns detailing the specific event, event category, event start time and event stop time}

\item{event_col}{String. Column name of specific event name column. Defaults to first column.}

\item{event_type_col}{String. Column name of event category column. Defaults to second column.}

\item{start_col}{String. Column name of event start time column. Defaults to third column.}

\item{stop_col}{String. Column name of event stop time column. Defaults to fourth column.}

\item{unit}{String. Unit of time to label the x-axis.}

\item{plotly}{Boolean. If TRUE, function will return an interactive plotly object.}

\item{title}{String. Desired plot title.}

\item{title_font}{Numeric. Size of title font.}

\item{x_font}{Numeric. Size of font on the x-axis.}

\item{y_font}{Numeric. Size of font on the y-axis}

\item{na.rm}{Boolean. If TRUE, events with incomplete time data will be removed.}
}
\value{
plot
}
\description{
\code{timelinePlot()} is used to generate a patient timeline plot to view longitudinal data.
}
\examples{
# Basic example
df <- data.frame(Event = c("Diagnosis","Death","Medication: Carboplatin","Medication: Carboplatin",
                           "Imaging: Diagnostic","Imaging: Surveillance","Surgery Biopsy: Peritoneum, NOS"),
                 EventType = c("Clinical Time Point","Clinical Time Point","Medication",
                               "Medication","Imaging","Imaging","Surgery Biopsy"),
                 AgeAtEventStart = c(69.085,71.507,69.189,69.441,69.142,70.134,69.085),
                 AgeAtEventStop = c(69.085,71.507,69.326,69.647,69.142,70.134,69.085))

timelinePlot(data = df, event_col = "Event", event_type_col = "EventType",
             start_col = "AgeAtEventStart", stop_col = "AgeAtEventStop", unit = "Years")

# Plotly Example
df <- data.frame(Event = c("Diagnosis","Death","Medication: Carboplatin","Medication: Carboplatin",
                           "Imaging: Diagnostic","Imaging: Surveillance","Surgery Biopsy: Peritoneum, NOS"),
                 EventType = c("Clinical Time Point","Clinical Time Point","Medication",
                               "Medication","Imaging","Imaging","Surgery Biopsy"),
                 AgeAtEventStart = c(69.085,71.507,69.189,69.441,69.142,70.134,69.085),
                 AgeAtEventStop = c(69.085,71.507,69.326,69.647,69.142,70.134,69.085),
                 Histology = c("84613 Serous Surface Papillary Carcinoma",NA,NA,NA,NA,NA,NA),
                 SpecimenSiteOfOrigin = c("Peritoneum, NOS",NA,NA,NA,NA,NA,NA),
                 CauseOfDeath = c(NA,"Probably Due to Cancer",NA,NA,NA,NA,NA),
                 MedicationTreatmentLine = c(NA,NA,"First Line","First Line",NA,NA,NA),
                 EvidenceOfLesion = c(NA,NA,NA,NA,TRUE,FALSE,NA),
                 SiteDiagnostic = c(NA,NA,NA,NA,NA,NA,TRUE))

timelinePlot(data = df, event_col = "Event", event_type_col = "EventType",
             start_col = "AgeAtEventStart", stop_col = "AgeAtEventStop", unit = "Years", plotly = TRUE)
}
