% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateDataTable}
\alias{updateDataTable}
\title{Update an dataTableOutput object. Can be used instead of renderDataTable}
\usage{
updateDataTable(session = NULL, id, value, app = getApp(session), ...)
}
\description{
Update an dataTableOutput object. Can be used instead of renderDataTable
}

