% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{updateDataTable}
\alias{updateDataTable}
\title{Update an dataTableOutput object. Can be used instead of renderDataTable}
\usage{
updateDataTable(session = NULL, id, value, app = getApp(session), ...)
}
\description{
Update an dataTableOutput object. Can be used instead of renderDataTable
}

