% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{updateImage}
\alias{updateImage}
\title{Update an output object. Can be used instead of renderImage}
\usage{
updateImage(session = NULL, id, value, app = getApp(session), ...)
}
\description{
Update an output object. Can be used instead of renderImage
}

