% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{updatePlot}
\alias{updatePlot}
\title{update an plotOutput object. Can be used instead of renderPlot.}
\usage{
updatePlot(session = NULL, id, expr, app = getApp(session),
  update.env = parent.frame(), quoted = FALSE)
}
\description{
update an plotOutput object. Can be used instead of renderPlot.
}

