% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updatePrint}
\alias{updatePrint}
\title{Update an textOutput object. Can be used instead of renderPrint}
\usage{
updatePrint(id, expr, app = getApp(), update.env = parent.frame())
}
\description{
Update an textOutput object. Can be used instead of renderPrint
}

