% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateTable}
\alias{updateTable}
\title{Update an tableOutput object. Can be used instead of renderTable}
\usage{
updateTable(id, expr, app = getApp(), update.env = parent.frame())
}
\description{
Update an tableOutput object. Can be used instead of renderTable
}

