% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateTable}
\alias{updateTable}
\title{Update an tableOutput object. Can be used instead of renderTable}
\usage{
updateTable(session = NULL, id, value, app = getApp(session), ...)
}
\description{
Update an tableOutput object. Can be used instead of renderTable
}

