% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateText}
\alias{updateText}
\title{Update an textOutput object. Can be used instead of renderText}
\usage{
updateText(id, expr, update.env = parent.frame(), app = getApp())
}
\description{
Update an textOutput object. Can be used instead of renderText
}

