% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRendered.r
\name{updateUI}
\alias{updateUI}
\title{Update an uiOutput object. Can be used instead of renderUI}
\usage{
updateUI(session, id, ui, app = getApp(session), ...)
}
\description{
Update an uiOutput object. Can be used instead of renderUI
}

