% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wasAceHotkeyPressed}
\alias{wasAceHotkeyPressed}
\title{Checks whether a button has been pressed again (internal function)}
\usage{
wasAceHotkeyPressed(keyId, value, app = getApp())
}
\description{
Checks whether a button has been pressed again (internal function)
}

